refactor: delete column
